DAOS-9379 [java] Adjust code to adapt to DAOS API change (daos_obj_generate_oid) in v2.0
