+++ src/client/java/hadoop-daos/src/main/java/io/daos/fs/hadoop/DaosFileSystem.java
@@ -377,7 +377,7 @@ public class DaosFileSystem extends FileSystem {
 
     daosFile.createNewFile(
             Constants.DAOS_MODLE,
-            DaosObjectType.OC_SX,
+            DaosObjectClass.OC_SX,
             this.chunkSize,
             true);
